OSS- IIT Dharwad
Open Student Society(OSS) is a community for students to join together to develop, innovate, share and grow together as a community. It is important to note that OSS is not a club, rather a community that welcomes students, from all branches and walks of life, to participate and contribute.